Add Rollbar::Middleware::Rack
Convert `Rollbar::Middleware::Rack` to be a class instead of a module,
so it can be added to the Rack middleware chain with `use
Rollbar::Middleware::Rack`.

Also, move the code from `Rollbar::Middleware::Sinatra` to
`Rollbar::Middleware::Rack` since Sinatra behavior is a subset of Rack
behavior.
